requirements-dev.txt
pyhamcrest
mock
pytest

-r requirements.txt
